- We fix the bobbing - Correct lot of warning - Movement now working with win size anymore - Add some sprites - Correct background imaging

+++ srcs/collision.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   collision.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fousse <fousse@student.42.fr>              +#+  +:+       +#+        */
+/*   By: gcollet <gcollet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/31 18:55:56 by fousse            #+#    #+#             */
-/*   Updated: 2022/01/06 11:46:16 by fousse           ###   ########.fr       */
+/*   Updated: 2022/01/10 17:01:05 by gcollet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,13 +27,13 @@ int	check_collision(int x, int y, int size, int map)
 	char		*tiles;
 
 	tiles = g_game.map.tiles;
-	if (tiles[(x - size) / TILE_SIZE + (y - size) / TILE_SIZE * map] == M_WALL)
+	if (tiles[(x - size) / (int)TILE_SIZE + (y - size) / (int)TILE_SIZE * map] == M_WALL)
 		return (1);
-	if (tiles[(x - size) / TILE_SIZE + (y + size) / TILE_SIZE * map] == M_WALL)
+	if (tiles[(x - size) / (int)TILE_SIZE + (y + size) / (int)TILE_SIZE * map] == M_WALL)
 		return (1);
-	if (tiles[(x + size) / TILE_SIZE + (y - size) / TILE_SIZE * map] == M_WALL)
+	if (tiles[(x + size) / (int)TILE_SIZE + (y - size) / (int)TILE_SIZE * map] == M_WALL)
 		return (1);
-	if (tiles[(x + size) / TILE_SIZE + (y + size) / TILE_SIZE * map] == M_WALL)
+	if (tiles[(x + size) / (int)TILE_SIZE + (y + size) / (int)TILE_SIZE * map] == M_WALL)
 		return (1);
 	return (0);
 }
